+++ generator/page_generators/downloads.js
@@ -11,10 +11,11 @@
 const Page = require("../Page.js").Page;
 const html = require("../html.js");
 
-let page = new Page("404", "error404.html");
+let page = new Page("Downloads", "downloads.html");
 
-page.h1("<span style=\"color: #ee1111;\">" + html.santitize("<Error|W404>") + " The page you were looking for does not exist</span>");
+page.h1("Downloads");
 
+page.paragraph("No releases have been made yet, but you can download the source code " + html.link("here", "https://github.com/PCIT-Project/PCIT-CPP") + " and build it yourself.");
 
 page.generate();