+++ generator/page_generators/home.js
@@ -1,27 +1,51 @@
+//////////////////////////////////////////////////////////////////////////////////////////////////
+//                                                                                              //
+// Part of pcit-project.github.io, under the Apache License v2.0 with LLVM and PCIT exceptions. //
+// You may not use this file except in compliance with the License.                             //
+// See `https://github.com/PCIT-Project/PCIT-CPP/blob/main/LICENSE`for info.                    //
+//                                                                                              //
+//////////////////////////////////////////////////////////////////////////////////////////////////
+
+
+
 const Page = require("../Page.js").Page;
 const html = require("../html.js");
 
-let home = new Page("Home", "home.html");
+let page = new Page("Home", "home.html");
 
-home.h1("PCIT Project");
-home.paragraph("Panther Compiler Infrastructure and Toolchain, and the home of the Panther programming language. PCIT (pronounced \"P-Kit\") relies on the " + html.link("https://github.com/llvm/llvm-project", "LLVM-Project") + ", but that is not guaranteed forever. It is written entirely in C++, but a self-hosted version (written entirely in Panther) is planned.");
+page.h1("What is PCIT Project?");
+page.paragraph("The PCIT Project is Panther Compiler Infrastructure and Toolchain, and the home of the " + "Panther programming language" + ". PCIT (pronounced \"P-Kit\") relies on the " + html.link("LLVM Project", "https://github.com/llvm/llvm-project") + ", but the goal is to eventually be self-reliant. It is written entirely in C++, but a self-hosted version (written entirely in Panther) is planned.");
+
+page.paragraph("The goal of the PCIT Project is to create software that helps developers create native software, and makes it easier to do so. The simple things should be simple, and the complicated things should be well documented and easy to use.");
+
+page.paragraph("Not only can all tools be used as a standalone program, all tools are designed to be embedded in your own projects as well.");
+
+page.begin_info();
+page.h2("Important Note", "padding-top: 0px;");
+page.paragraph("PCIT is very early development. Please do not use any tools to create anything yet as drastic changes may be made.");
+page.end_info();
+
+page.h2("Tools");
+
+page.h3("Panther / pthr");
+page.paragraph("Statically typed, compiled, high-performance, general-purpose programming language. Designed to help programmers write good/fast code that, without getting in the way. Seamless interoperability with C and (hopefully) C++. All this while being enjoyable to use. See more " + html.link("here", "./Panther.html") + ".");
+
+page.h3("PIR");
+page.paragraph("Pronounced \"P I R\". Compiler IR and SSA-based optimizing back-end.");
+
+page.h3("PLNK");
+page.paragraph("Pronounced \"plink\". Linker that aims to make use between platforms (including cross-compilation and linking against libc) as seamless as possible.");
 
 
-home.h2("Tools");
-home.paragraph("All tools can be used as stand-alone programs or as libraries.");
+page.h3("Other compilers");
+page.paragraph("Will also be able to compile C and hopefully C++ code. The main intent is for this to allow seamless interoperability with Panther, but these frontends will be able to be used stand-alone.");
 
-home.h3("Panther / pthr");
-home.paragraph("Statically typed, compiled, high-performance, general-purpose programming language. See more " + html.link("./Panther.html", "here") + ".");
+page.h2("Licensing");
+page.paragraph("All of the code written for the PCIT Project is licensed under the Apache License v2.0 with LLVM and PCIT exceptions. See " + html.link("the license", "https://github.com/PCIT-Project/PCIT-CPP/blob/main/LICENSE") + " for more information.");
 
-home.h3("PIR");
-home.paragraph("Pronounced \"P I R\". Compiler IR / optimizing back-end.");
-
-home.h3("PLNK");
-home.paragraph("Pronounced \"plink\". Linker that aims to make use between platforms (including cross-compilation) as seamless as possible.");
+page.h2("Getting involved");
+page.paragraph("The PCIT Project is not accepting any volunteers at the moment. However, at some point the in the future new contributors will be welcomed. Check back here to if we are accepting contributions. For now, please star " + html.link("the GitHub repository", "https://github.com/PCIT-Project/PCIT-CPP") + " to show you're interested in the project.");
 
 
-home.h3("Other compilers");
-home.paragraph("Will also be able to compile C and hopefully C++ code. The main intent is for this to allow seamless interoperability with Panther, but these frontends will be able to be used stand-alone.");
-
-home.generate();
+page.generate(true);